Re-implement AbuseFilter integration in a scalable way
The previous implementation of the sfs-confidence variable was not very
scalable as it made an API request for each IP address. This was also
quesitonable in terms of privacy since it exposed user IP addresses to a
third-party site.

So intead implement a sfs-blocked variable which simply indicates
whether a user is blocked on the list or not.

Change-Id: I7408fb6b8ace14b487a0c49ffe28ee347476d9ef